+++ CopilotKit/packages/react-ui/src/components/chat/Button.tsx
@@ -3,8 +3,7 @@ import { useChatContext } from "./ChatContext";
 
 export const Button = ({ open, setOpen }: ButtonProps) => {
   const context = useChatContext();
-  // To ensure that the mouse handler fires even when the button is scaled down
-  // we wrap the button in a div and attach the handler to the div
+
   return (
     <div onClick={() => setOpen(!open)}>
       <button